Add new checks for the ui-allinone
