--- tests/e2e/allinone/Makefile
@@ -1,5 +1,0 @@
-render-e2e-tests-allinone: load-assert-job
-	./tests/e2e/allinone/render.sh
-
-run-e2e-tests-allinone: TEST_SUITE_NAME=allinone
-run-e2e-tests-allinone: run-suite-tests
