--- tests/e2e/allinone/allinone-ingress/00-assert.yaml
@@ -1,7 +1,0 @@
-# Assert there is a running Jaeger instance
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: my-jaeger
-spec:
-  replicas: 1
