--- tests/e2e/allinone/allinone-ingress/00-install.yaml
@@ -1,14 +1,0 @@
-# Install the Jaeger intance
-apiVersion: jaegertracing.io/v1
-kind: Jaeger
-metadata:
-  name: my-jaeger
-spec:
-  strategy: allinone
-  allInOne:
-    options:
-      log-level:  "debug"
-      memory.max-traces: 10000
-  ingress:
-    enabled: true
-    security: "none"
