--- tests/e2e/allinone/allinone-ingress/README.md
@@ -1,5 +1,0 @@
-# All in One - Ingress
-## What is this test case testing?
-
-This test case tests if the parameter `ingress` works properly creating a route
-to access the Jaeger UI from outside the cluster.
