--- tests/e2e/allinone/allinone-smoke/00-assert.yaml
@@ -1,6 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: my-jaeger
-spec:
-  replicas: 1
