--- tests/e2e/allinone/allinone-smoke/README.md
@@ -1,4 +1,0 @@
-# All in One - Smoke
-## What is this test case testing?
-
-Smoke test for the All in One deployment strategy.
