+++ tests/e2e/ui/allinone/00-assert.yaml
