+++ tests/e2e/ui/allinone/02-install.yaml
@@ -1,18 +1,11 @@
+# Enable tracking.gaID
 apiVersion: jaegertracing.io/v1
 kind: Jaeger
 metadata:
   name: all-in-one-with-ui-config
-  anotations:
-    nginx.ingress.kubernetes.io/ssl-redirect: "false"
 spec:
   strategy: allinone
-  allInOne:
-    options:
-      query.base-path: "/jaeger"
   ui:
     options:
       tracking:
         gaID: "MyTrackingId"
-  ingress:
-    enabled: true
-    security: "none"
