+++ tests/e2e/ui/allinone/curl.yaml.template
@@ -1,0 +1,5 @@
+apiVersion: kuttl.dev/v1beta1
+kind: TestStep
+commands:
+  - command: "./ensure-ingress-host.sh"
+  - script: "{{ .Env.ASSERT_HTTP_CODE_PROGRAM }} {{ .Env.URL }} {{ .Env.EXPECTED_CODE }}"
