--- tests/e2e/ui/ui-definition/02-assert.yaml
@@ -1,6 +1,0 @@
-apiVersion: batch/v1
-kind: Job
-metadata:
-  name: test-ui-config
-status:
-  succeeded: 1
